+++ tests/fixtures/nuxt-content-components-without-module/components/content/Alert.vue
@@ -1,0 +1,3 @@
+<template>
+  <div class="alert"><slot /></div>
+</template>