+++ tests/fixtures/nuxt-global-components/components/Baz.global.vue
@@ -1,0 +1,3 @@
+<template>
+  <p>global baz</p>
+</template>